tools/export_onnx/requirements.txt
torch
nemo_toolkit[asr]
onnx
onnxruntime
onnx-simplifier
